Revert "Move loadStdLibs to package manager"
This reverts commit 47a79eb1640651b8f8c7c5bc498b360e88c320ed.
