+++ app/Main.hs
@@ -3,6 +3,11 @@ module Main (main) where
 ------------------------------------------------------------------------------
 
 import qualified Data.ByteString as BS
+import qualified Data.ByteString.Base64 as BS
+import GitHub (github)
+import qualified GitHub
+import qualified GitHub.Auth as GitHub
+import qualified GitHub.Endpoints.Repos.Contents as GitHub
 import qualified Juvix.Backends.LLVM as LLVM
 import qualified Juvix.Backends.Michelson as Michelson
 import qualified Juvix.Backends.Plonk as Plonk
@@ -17,7 +22,6 @@ import Text.Pretty.Simple
 import Text.Pretty.Simple (pPrint)
 import Text.PrettyPrint.ANSI.Leijen (putDoc)
 import Version (infoVersionRepo, progNameVersionTag)
-import Juvix.PackageManager.StdLib
 
 ------------------------------------------------------------------------------
 -- Run commands
@@ -90,6 +94,34 @@ runCmd' ::
   Pipeline.Pipeline ()
 runCmd' fin b f = liftIO (readFile fin) >>= f b >>= liftIO . pPrint
 
+installStdLibs :: IO ()
+installStdLibs = do
+    -- TODO: If found locally, do not fetch (unless flag is set to force fetch)
+    getContents "stdlib"
+    where
+      getJuvixHome = (<> "/.juvix/") <$> getHomeDirectory
+      createDir p = do
+        d <- getJuvixHome 
+        createDirectoryIfMissing True (d <> p) 
+
+
+      getContents :: Text -> IO ()
+      getContents path = do
+        stdlibsR <- github (GitHub.OAuth "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") 
+          -- This key just has public repo read access
+          $ GitHub.contentsForR "anoma" "juvix" path Nothing
+
+      case stdlibsR of
+        Left err -> pPrint err
+        Right (GitHub.ContentDirectory stdlibs) -> do
+          createDir (toS path)
+          traverse_ (\stdlib -> getContents (GitHub.contentPath $ GitHub.contentItemInfo stdlib)) stdlibs
+        Right (GitHub.ContentFile file) -> do
+          let content = GitHub.contentFileContent file
+          let path = GitHub.contentPath $ GitHub.contentFileInfo file
+          localJuvix <- getJuvixHome
+          BS.writeFile (localJuvix <> (toS path)) (BS.decodeLenient $ encodeUtf8 content)
+
 ------------------------------------------------------------------------------
 -- Main
 ------------------------------------------------------------------------------
@@ -100,6 +132,6 @@ main = do
   home <- getHomeDirectory
   let ctx = Context pwd home
   progVersion <- progNameVersionTag
-  loadStdLibs
+  installStdLibs
   let opts = info (options ctx <**> helper) (fullDesc <> headerDoc (Just progVersion))
   run ctx =<< execParser opts
