--- library/Test/PackageManager/.gitignore
@@ -1,2 +1,0 @@
-.stack-work/
-*~
