--- library/Test/PackageManager/Setup.hs
@@ -1,3 +1,0 @@
-import Distribution.Simple
-
-main = defaultMain
