--- library/Test/PackageManager/test/Spec.hs
@@ -1,4 +1,0 @@
-import Juvix.Library
-
-main :: IO ()
-main = putStrLn ("Test suite not yet implemented" :: [Char])
