+++ stack.yaml
@@ -17,7 +17,6 @@ packages:
 - ./library/Context
 - ./library/Sexp
 - ./library/Test/DataStructures
-- ./library/Test/PackageManager
 
 extra-deps:
